+++ src/main/java/com/github/manasmods/tensura/gui/screen/Chapter1Screen.java
@@ -1,16 +1,10 @@
 package com.github.manasmods.tensura.gui.screen;
 
 import com.github.manasmods.tensura.Tensura;
-import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.network.chat.TextComponent;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.stats.Stats;
-import net.minecraft.world.InteractionHand;
-import net.minecraft.world.InteractionResultHolder;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.Level;
 
 public class Chapter1Screen extends Screen {
 
@@ -86,16 +80,5 @@ public class Chapter1Screen extends Screen {
 
     public Chapter1Screen(TextComponent chapter1ItemGui, Player pPlayer) {
         super(chapter1ItemGui, pPlayer);
-    }
-
-    public InteractionResultHolder<ItemStack> use(Level pLevel, Player pPlayer, InteractionHand pHand) {
-        ItemStack itemstack = pPlayer.getItemInHand(pHand);
-        pPlayer.openItemGui(itemstack, pHand);
-        pPlayer.awardStat(Stats.ITEM_USED.get(this));
-        return InteractionResultHolder.sidedSuccess(itemstack, pLevel.isClientSide());
-        if (pLevel.isClientSide) {
-            Minecraft.getInstance().setScreen(new Chapter1Screen(new TextComponent("chapter1ItemGui"), pPlayer));
-
-        }
     }
 }
